requirements.txt
ultralytics
streamlit
Pillow
numpy
opencv-python-headless